// src/generated/user.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.7.1
//   protoc               v3.20.3
// source: user.proto

/* eslint-disable */
import { GrpcMethod, GrpcStreamMethod } from "@nestjs/microservices";
import { Observable } from "rxjs";

export const protobufPackage = "user";

export interface Empty {
}

export interface GetUserRequest {
  userId: string;
  requesterRole: string;
}

export interface User {
  id: string;
  name: string;
  email: string;
  isBlocked: boolean;
  followerCount: number;
}

export interface UserResponse {
  user: User | undefined;
}

export interface UserListResponse {
  users: User[];
}

export const USER_PACKAGE_NAME = "user";

export interface UserServiceClient {
  getUser(request: GetUserRequest): Observable<UserResponse>;

  getAllUsers(request: Empty): Observable<UserListResponse>;

  blockUser(request: GetUserRequest): Observable<UserResponse>;

  getTopInfluencer(request: Empty): Observable<UserListResponse>;
}

export interface UserServiceController {
  getUser(request: GetUserRequest): Promise<UserResponse> | Observable<UserResponse> | UserResponse;

  getAllUsers(request: Empty): Promise<UserListResponse> | Observable<UserListResponse> | UserListResponse;

  blockUser(request: GetUserRequest): Promise<UserResponse> | Observable<UserResponse> | UserResponse;

  getTopInfluencer(request: Empty): Promise<UserResponse> | Observable<UserResponse> | UserResponse;
}

export function UserServiceControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = ["getUser", "getAllUsers", "blockUser", "getTopInfluencer"];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcMethod("UserService", method)(constructor.prototype[method], method, descriptor);
    }
    const grpcStreamMethods: string[] = [];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcStreamMethod("UserService", method)(constructor.prototype[method], method, descriptor);
    }
  };
}

export const USER_SERVICE_NAME = "UserService";