// src/generated/post.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.7.2
//   protoc               v3.20.3
// source: post.proto

/* eslint-disable */
import { GrpcMethod, GrpcStreamMethod } from "@nestjs/microservices";
import { Observable } from "rxjs";

export const protobufPackage = "post";

export interface Empty {
}

export interface Post {
  id: string;
  userId: string;
  content: string;
  media: string[];
  visibility: string;
  deleted: boolean;
  moderated: boolean;
  keywords: string[];
  createdAt: string;
  updatedAt: string;
  isReported: boolean;
  reportReason: string;
}

export interface PostList {
  posts: Post[];
}

export interface PostResponse {
  post: Post | undefined;
}

export interface FlagPostRequest {
  postId: string;
  reason: string;
}

export interface PostIdRequest {
  postId: string;
}

export const POST_PACKAGE_NAME = "post";

export interface PostServiceClient {
  AllPosts(request: Empty): Observable<PostList>;

  reportedPosts(request: Empty): Observable<PostList>;

  flagPost(request: FlagPostRequest): Observable<PostResponse>;

  adminDeletePost(request: PostIdRequest): Observable<PostResponse>;
}

export interface PostServiceController {
  AllPosts(request: Empty): Promise<PostList> | Observable<PostList> | PostList;

  reportedPosts(request: Empty): Promise<PostList> | Observable<PostList> | PostList;

  flagPost(request: FlagPostRequest): Promise<PostResponse> | Observable<PostResponse> | PostResponse;

  adminDeletePost(request: PostIdRequest): Promise<PostResponse> | Observable<PostResponse> | PostResponse;
}

export function PostServiceControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = ["allPosts", "reportedPosts", "flagPost", "adminDeletePost"];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcMethod("PostService", method)(constructor.prototype[method], method, descriptor);
    }
    const grpcStreamMethods: string[] = [];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcStreamMethod("PostService", method)(constructor.prototype[method], method, descriptor);
    }
  };
}

export const POST_SERVICE_NAME = "PostService";